# Speeding up the tests by reducing the number of K.eval().