### Summary

In this PR, I adapted the change in the commit e40ffd2 to other tests. Giving interesting speedups. The numbers here were obtained on my laptop, using tensorflow.

test_random_uniform:
5.92s -> 0.18s

test_random_binomial:
5.95s -> 0.23s

### Related Issues

### PR Overview

* I kept the number of samples generated identical.
* I converted the list to a numpy array directly to avoid multiple conversions `list` -> `ndarray`.

- [ ] This PR requires new unit tests [y/n] (make sure tests are included)
- [ ] This PR requires to update the documentation [y/n] (make sure the docs are up-to-date)
- [x] This PR is backwards compatible [y/n]
- [ ] This PR changes the current API [y/n] (all API changes need to be approved by fchollet)